backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.39.0
supabase==2.3.0
python-dotenv==1.0.0
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
python-magic==0.4.27
markdown==3.5.1
gunicorn==21.2.0
